+++ GamePrograming/Application/Application.vcxproj
@@ -204,7 +204,7 @@
     <ClInclude Include="Source\DirectX\Texture.h" />
     <ClInclude Include="Source\Game\Controller.h" />
     <ClInclude Include="Source\Game\Field.h" />
-    <ClInclude Include="Source\Game\FieldObject.h" />
+    <ClInclude Include="Source\Game\ThrowObject.h" />
     <ClInclude Include="Source\Game\GameObject.h" />
     <ClInclude Include="Source\Game\GameScene.h" />
     <ClInclude Include="Source\Game\GameSystem.h" />
@@ -222,7 +222,7 @@
     <ClCompile Include="Source\DirectX\Texture.cpp" />
     <ClCompile Include="Source\Game\Controller.cpp" />
     <ClCompile Include="Source\Game\Field.cpp" />
-    <ClCompile Include="Source\Game\FieldObject.cpp" />
+    <ClCompile Include="Source\Game\ThrowObject.cpp" />
     <ClCompile Include="Source\Game\GameObject.cpp" />
     <ClCompile Include="Source\Game\GameScene.cpp" />
     <ClCompile Include="Source\Game\GameSystem.cpp" />
