+++ GamePrograming/Application/x64/Debug/Application.exe.recipe
@@ -2,7 +2,7 @@
 <Project>
   <ProjectOutputs>
     <ProjectOutput>
-      <FullPath>C:\Users\ths30712\source\HEW\UpTempoProject\GamePrograming\x64\Debug\Application.exe</FullPath>
+      <FullPath>Z:\Documents\GitHub\UpTempoProject\GamePrograming\x64\Debug\Application.exe</FullPath>
     </ProjectOutput>
   </ProjectOutputs>
   <ContentFiles />
