+++ GamePrograming/Application/x64/Debug/Application.log
@@ -1,2 +1,1 @@
-﻿  Field.cpp
-  Application.vcxproj -> C:\Users\ths30712\source\HEW\UpTempoProject\GamePrograming\x64\Debug\Application.exe
+﻿  Application.vcxproj -> Z:\Documents\GitHub\UpTempoProject\GamePrograming\x64\Debug\Application.exe
