+++ GamePrograming/Application/x64/Debug/Application.tlog/Application.lastbuildstate
@@ -1,2 +1,2 @@
-PlatformToolSet=v143:VCToolArchitecture=Native64Bit:VCToolsVersion=14.41.34120:TargetPlatformVersion=10.0.22000.0:
-Debug|x64|C:\Users\ths30712\source\HEW\UpTempoProject\GamePrograming\|
+PlatformToolSet=v143:VCToolArchitecture=NativeARM64:VCToolsVersion=14.41.34120:TargetPlatformVersion=10.0.22621.0:
+Debug|x64|Z:\Documents\GitHub\UpTempoProject\GamePrograming\|
