+++ GamePrograming/Application/x64/Debug/Application.tlog/Cl.items.tlog
@@ -1,14 +1,14 @@
-C:\Users\ths30712\source\HEW\UpTempoProject\GamePrograming\Application\Application.cpp;C:\Users\ths30712\source\HEW\UpTempoProject\GamePrograming\Application\x64\Debug\Application.obj
-C:\Users\ths30712\source\HEW\UpTempoProject\GamePrograming\Application\Source\DirectX\Audio.cpp;C:\Users\ths30712\source\HEW\UpTempoProject\GamePrograming\Application\x64\Debug\Audio.obj
-C:\Users\ths30712\source\HEW\UpTempoProject\GamePrograming\Application\Source\DirectX\Direct3D.cpp;C:\Users\ths30712\source\HEW\UpTempoProject\GamePrograming\Application\x64\Debug\Direct3D.obj
-C:\Users\ths30712\source\HEW\UpTempoProject\GamePrograming\Application\Source\DirectX\Input.cpp;C:\Users\ths30712\source\HEW\UpTempoProject\GamePrograming\Application\x64\Debug\Input.obj
-C:\Users\ths30712\source\HEW\UpTempoProject\GamePrograming\Application\Source\DirectX\Texture.cpp;C:\Users\ths30712\source\HEW\UpTempoProject\GamePrograming\Application\x64\Debug\Texture.obj
-C:\Users\ths30712\source\HEW\UpTempoProject\GamePrograming\Application\Source\Game\Controller.cpp;C:\Users\ths30712\source\HEW\UpTempoProject\GamePrograming\Application\x64\Debug\Controller.obj
-C:\Users\ths30712\source\HEW\UpTempoProject\GamePrograming\Application\Source\Game\Field.cpp;C:\Users\ths30712\source\HEW\UpTempoProject\GamePrograming\Application\x64\Debug\Field.obj
-C:\Users\ths30712\source\HEW\UpTempoProject\GamePrograming\Application\Source\Game\FieldObject.cpp;C:\Users\ths30712\source\HEW\UpTempoProject\GamePrograming\Application\x64\Debug\FieldObject.obj
-C:\Users\ths30712\source\HEW\UpTempoProject\GamePrograming\Application\Source\Game\GameObject.cpp;C:\Users\ths30712\source\HEW\UpTempoProject\GamePrograming\Application\x64\Debug\GameObject.obj
-C:\Users\ths30712\source\HEW\UpTempoProject\GamePrograming\Application\Source\Game\GameScene.cpp;C:\Users\ths30712\source\HEW\UpTempoProject\GamePrograming\Application\x64\Debug\GameScene.obj
-C:\Users\ths30712\source\HEW\UpTempoProject\GamePrograming\Application\Source\Game\GameSystem.cpp;C:\Users\ths30712\source\HEW\UpTempoProject\GamePrograming\Application\x64\Debug\GameSystem.obj
-C:\Users\ths30712\source\HEW\UpTempoProject\GamePrograming\Application\Source\Game\Ground.cpp;C:\Users\ths30712\source\HEW\UpTempoProject\GamePrograming\Application\x64\Debug\Ground.obj
-C:\Users\ths30712\source\HEW\UpTempoProject\GamePrograming\Application\Source\Game\Physics.cpp;C:\Users\ths30712\source\HEW\UpTempoProject\GamePrograming\Application\x64\Debug\Physics.obj
-C:\Users\ths30712\source\HEW\UpTempoProject\GamePrograming\Application\Source\Game\Player.cpp;C:\Users\ths30712\source\HEW\UpTempoProject\GamePrograming\Application\x64\Debug\Player.obj
+Z:\Documents\GitHub\UpTempoProject\GamePrograming\Application\Application.cpp;Z:\Documents\GitHub\UpTempoProject\GamePrograming\Application\x64\Debug\Application.obj
+Z:\Documents\GitHub\UpTempoProject\GamePrograming\Application\Source\DirectX\Audio.cpp;Z:\Documents\GitHub\UpTempoProject\GamePrograming\Application\x64\Debug\Audio.obj
+Z:\Documents\GitHub\UpTempoProject\GamePrograming\Application\Source\DirectX\Direct3D.cpp;Z:\Documents\GitHub\UpTempoProject\GamePrograming\Application\x64\Debug\Direct3D.obj
+Z:\Documents\GitHub\UpTempoProject\GamePrograming\Application\Source\DirectX\Input.cpp;Z:\Documents\GitHub\UpTempoProject\GamePrograming\Application\x64\Debug\Input.obj
+Z:\Documents\GitHub\UpTempoProject\GamePrograming\Application\Source\DirectX\Texture.cpp;Z:\Documents\GitHub\UpTempoProject\GamePrograming\Application\x64\Debug\Texture.obj
+Z:\Documents\GitHub\UpTempoProject\GamePrograming\Application\Source\Game\Controller.cpp;Z:\Documents\GitHub\UpTempoProject\GamePrograming\Application\x64\Debug\Controller.obj
+Z:\Documents\GitHub\UpTempoProject\GamePrograming\Application\Source\Game\Field.cpp;Z:\Documents\GitHub\UpTempoProject\GamePrograming\Application\x64\Debug\Field.obj
+Z:\Documents\GitHub\UpTempoProject\GamePrograming\Application\Source\Game\ThrowObject.cpp;Z:\Documents\GitHub\UpTempoProject\GamePrograming\Application\x64\Debug\ThrowObject.obj
+Z:\Documents\GitHub\UpTempoProject\GamePrograming\Application\Source\Game\GameObject.cpp;Z:\Documents\GitHub\UpTempoProject\GamePrograming\Application\x64\Debug\GameObject.obj
+Z:\Documents\GitHub\UpTempoProject\GamePrograming\Application\Source\Game\GameScene.cpp;Z:\Documents\GitHub\UpTempoProject\GamePrograming\Application\x64\Debug\GameScene.obj
+Z:\Documents\GitHub\UpTempoProject\GamePrograming\Application\Source\Game\GameSystem.cpp;Z:\Documents\GitHub\UpTempoProject\GamePrograming\Application\x64\Debug\GameSystem.obj
+Z:\Documents\GitHub\UpTempoProject\GamePrograming\Application\Source\Game\Ground.cpp;Z:\Documents\GitHub\UpTempoProject\GamePrograming\Application\x64\Debug\Ground.obj
+Z:\Documents\GitHub\UpTempoProject\GamePrograming\Application\Source\Game\Physics.cpp;Z:\Documents\GitHub\UpTempoProject\GamePrograming\Application\x64\Debug\Physics.obj
+Z:\Documents\GitHub\UpTempoProject\GamePrograming\Application\Source\Game\Player.cpp;Z:\Documents\GitHub\UpTempoProject\GamePrograming\Application\x64\Debug\Player.obj
