+++ GamePrograming/Application/x64/Debug/Application.tlog/link.secondary.1.tlog
@@ -1,2 +1,2 @@
-^C:\USERS\THS30712\SOURCE\HEW\UPTEMPOPROJECT\GAMEPROGRAMING\APPLICATION\X64\DEBUG\APPLICATION.OBJ|C:\USERS\THS30712\SOURCE\HEW\UPTEMPOPROJECT\GAMEPROGRAMING\APPLICATION\X64\DEBUG\APPLICATION.RES|C:\USERS\THS30712\SOURCE\HEW\UPTEMPOPROJECT\GAMEPROGRAMING\APPLICATION\X64\DEBUG\AUDIO.OBJ|C:\USERS\THS30712\SOURCE\HEW\UPTEMPOPROJECT\GAMEPROGRAMING\APPLICATION\X64\DEBUG\CONTROLLER.OBJ|C:\USERS\THS30712\SOURCE\HEW\UPTEMPOPROJECT\GAMEPROGRAMING\APPLICATION\X64\DEBUG\DIRECT3D.OBJ|C:\USERS\THS30712\SOURCE\HEW\UPTEMPOPROJECT\GAMEPROGRAMING\APPLICATION\X64\DEBUG\FIELD.OBJ|C:\USERS\THS30712\SOURCE\HEW\UPTEMPOPROJECT\GAMEPROGRAMING\APPLICATION\X64\DEBUG\FIELDOBJECT.OBJ|C:\USERS\THS30712\SOURCE\HEW\UPTEMPOPROJECT\GAMEPROGRAMING\APPLICATION\X64\DEBUG\GAMEOBJECT.OBJ|C:\USERS\THS30712\SOURCE\HEW\UPTEMPOPROJECT\GAMEPROGRAMING\APPLICATION\X64\DEBUG\GAMESCENE.OBJ|C:\USERS\THS30712\SOURCE\HEW\UPTEMPOPROJECT\GAMEPROGRAMING\APPLICATION\X64\DEBUG\GAMESYSTEM.OBJ|C:\USERS\THS30712\SOURCE\HEW\UPTEMPOPROJECT\GAMEPROGRAMING\APPLICATION\X64\DEBUG\GROUND.OBJ|C:\USERS\THS30712\SOURCE\HEW\UPTEMPOPROJECT\GAMEPROGRAMING\APPLICATION\X64\DEBUG\INPUT.OBJ|C:\USERS\THS30712\SOURCE\HEW\UPTEMPOPROJECT\GAMEPROGRAMING\APPLICATION\X64\DEBUG\PHYSICS.OBJ|C:\USERS\THS30712\SOURCE\HEW\UPTEMPOPROJECT\GAMEPROGRAMING\APPLICATION\X64\DEBUG\PLAYER.OBJ|C:\USERS\THS30712\SOURCE\HEW\UPTEMPOPROJECT\GAMEPROGRAMING\APPLICATION\X64\DEBUG\TEXTURE.OBJ
-C:\Users\ths30712\source\HEW\UpTempoProject\GamePrograming\Application\x64\Debug\Application.ilk
+^Z:\DOCUMENTS\GITHUB\UPTEMPOPROJECT\GAMEPROGRAMING\APPLICATION\X64\DEBUG\APPLICATION.OBJ|Z:\DOCUMENTS\GITHUB\UPTEMPOPROJECT\GAMEPROGRAMING\APPLICATION\X64\DEBUG\APPLICATION.RES|Z:\DOCUMENTS\GITHUB\UPTEMPOPROJECT\GAMEPROGRAMING\APPLICATION\X64\DEBUG\AUDIO.OBJ|Z:\DOCUMENTS\GITHUB\UPTEMPOPROJECT\GAMEPROGRAMING\APPLICATION\X64\DEBUG\CONTROLLER.OBJ|Z:\DOCUMENTS\GITHUB\UPTEMPOPROJECT\GAMEPROGRAMING\APPLICATION\X64\DEBUG\DIRECT3D.OBJ|Z:\DOCUMENTS\GITHUB\UPTEMPOPROJECT\GAMEPROGRAMING\APPLICATION\X64\DEBUG\FIELD.OBJ|Z:\DOCUMENTS\GITHUB\UPTEMPOPROJECT\GAMEPROGRAMING\APPLICATION\X64\DEBUG\GAMEOBJECT.OBJ|Z:\DOCUMENTS\GITHUB\UPTEMPOPROJECT\GAMEPROGRAMING\APPLICATION\X64\DEBUG\GAMESCENE.OBJ|Z:\DOCUMENTS\GITHUB\UPTEMPOPROJECT\GAMEPROGRAMING\APPLICATION\X64\DEBUG\GAMESYSTEM.OBJ|Z:\DOCUMENTS\GITHUB\UPTEMPOPROJECT\GAMEPROGRAMING\APPLICATION\X64\DEBUG\GROUND.OBJ|Z:\DOCUMENTS\GITHUB\UPTEMPOPROJECT\GAMEPROGRAMING\APPLICATION\X64\DEBUG\INPUT.OBJ|Z:\DOCUMENTS\GITHUB\UPTEMPOPROJECT\GAMEPROGRAMING\APPLICATION\X64\DEBUG\PHYSICS.OBJ|Z:\DOCUMENTS\GITHUB\UPTEMPOPROJECT\GAMEPROGRAMING\APPLICATION\X64\DEBUG\PLAYER.OBJ|Z:\DOCUMENTS\GITHUB\UPTEMPOPROJECT\GAMEPROGRAMING\APPLICATION\X64\DEBUG\TEXTURE.OBJ|Z:\DOCUMENTS\GITHUB\UPTEMPOPROJECT\GAMEPROGRAMING\APPLICATION\X64\DEBUG\THROWOBJECT.OBJ
+Z:\Documents\GitHub\UpTempoProject\GamePrograming\Application\x64\Debug\Application.ilk
